✨ Weighted classification loss
It is riskier to predict a "yes" or "no" for a pair of checkboxes that
are "empty" or "both" than it is to predict "both" or "empty" for a
"yes" or "no" pair.

Assign a higher weight to "yes" and "no" options, so that incorrect
predictions of "yes" or "no" are penalized more severely than incorrect
predictions of "both" or "empty".

[ch 97634] https://app.shortcut.com/diwoblood/story/97634/use-weighted-classification-loss-to-penalize-false-yes-and-no