Create detect_and_export.py
Added a class for object detection using the YOLOV7 model that can export detections as txt, json, or csv files. The class can make predictions from an image, video, or dataset of images. All detection functions return a list of dictionaries of detected object classes and counts for each image. The class preprocesses the input image for input to the YOLOv7 model and runs object detection on the image to return a dictionary of the detected objects and their counts. The detected objects can be exported to txt, json, or csv files. The class also runs the YOLOv7 model on a dataset of images and a video, returning a list of dictionaries of detected object classes and counts for each image. The class is made with love by https://github.com/nachovoss and is free to use and modify under the MIT license. Thanks to the awesome developers who made YOLOv7 possible.