✨ Remove using rectangular images as model inputs
In the previous experiments, using rectangular images as input data
during training to reduce model size and increase detection speed
downgrade the model performance significantly. Thus, remove using
rectangular images as input data during training. Convert the square
input shape into rectangular shape to reduce the model size and increase
the detection speed while converting a pretrained model into tflite
model.

Use the same resize scale in testing as used in training and validation
to resize input images. Previously the resize scale used in testing is
calculated based on the testing dataset, however the testing dataset is
much smaller than the training one. It is likely happen that the resize
scale calculated based on the testing dataset is not the same as the
one used in training and validation. Thus, hard code the resize scale in
testing script to be the same as the one used in training and validation.

Same for testing tflite models, use the same resize scale as used in
training and validation.

Use only the class with the highest classification confidence in the nms
algorithm.

To be compatible with `DVC` and `TensorFlow` used for converting a
`torch` model into a `tflite` one, the `conda` environment is updated.
Update the environment file accordingly.